LibUpdate_3.3.0
更新内容
-  美颜模块全面升级，改进磨皮算法，增加红润效果，改善脸型美化
-  手势识别模块升级，发布v2版手势识别模型，利用深度神经网络提高手势检出率，同时降低误检率
-  其他已知问题修复

+++ FULiveDemo/Faceunity/FaceUnity-SDK-iOS/Headers/funama.h
@@ -334,7 +334,7 @@ FUNAMA_API void fuSetQualityTradeoff(float quality);
 	0 means failure, general failure is due to invalid face info
 	other specific failure will print on the console
 */
-FUNAMA_API int fuGetFaceInfo(int face_id, char* name, double* pret, int num);
+FUNAMA_API int fuGetFaceInfo(int face_id, char* name, float* pret, int num);
 
 /**
 \brief Get SDK version string
